[wip][update] move ElfLoader to lib.rs
